Revert "stages: combine prepivot and storage stages"
This reverts commit 6f9e99da67a6fa2068c95494b036d7835e917ee2.
