+++ Makefile
@@ -5,6 +5,7 @@ PACKAGES = \
     src/config \
     src/exec \
     src/exec/stages \
+    src/exec/stages/prepivot \
     src/exec/stages/storage \
     src/exec/util \
     src/providers \
